Conflicts:
	scilab/modules/graphic_objects/Makefile.am
	scilab/modules/graphic_objects/Makefile.in
	scilab/modules/graphic_objects/src/c/graphic_objects.vcxproj.filters
	scilab/modules/graphic_objects/src/jni/DataLoader_wrap.c
	scilab/modules/graphic_objects/src/scripts/propertiesMap.properties
	scilab/modules/graphics/Makefile.am
	scilab/modules/graphics/Makefile.in
	scilab/modules/graphics/help/en_US/datatips/datatipSetDisplay.xml
	scilab/modules/graphics/help/en_US/datatips/datatipSetStyle.xml
	scilab/modules/graphics/macros/datatips/datatipCreate.sci
	scilab/modules/graphics/macros/datatips/datatipRemove.sci
	scilab/modules/graphics/macros/datatips/datatipSetInterp.sci
	scilab/modules/graphics/macros/datatips/datatipToggle.sci
	scilab/modules/graphics/src/c/getHandleProperty/GetHashTable.c
	scilab/modules/graphics/src/c/getHandleProperty/SetHashTable.c
	scilab/modules/graphics/src/c/getHandleProperty/getHandleProperty.h
	scilab/modules/graphics/src/c/getHandleProperty/setHandleProperty.h
	scilab/modules/gui/Makefile.am
	scilab/modules/gui/Makefile.in
	scilab/modules/localization/src/jni/Messages_wrap.c
	scilab/modules/types/src/jni/getScilabVariable_wrap.c

+++ scilab/modules/fileio/src/c/mgeti.h
@@ -7,7 +7,7 @@
  * This source file is licensed as described in the file COPYING, which
  * you should have received as part of this distribution.  The terms
  * are also available at
- * http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
+ * http://www.cecill.info/licences/Licence_CeCILL_V2.1-en.txt
  *
  */
 
@@ -35,7 +35,7 @@ int SWAP(char type[], int *fd);
  * @param char type[]
  * @param ierr
  */
-FILEIO_IMPEXP void C2F(mgeti) (int *fd, int *res, int *n, char type[], int *ierr);
+FILEIO_IMPEXP void C2F(mgeti) (int *fd, long long *res, int *n, char type[], int *ierr);
 
 #endif /* __MGETI_H__ */
