+++ scilab/modules/graphic_objects/src/cpp/Triangulator.cpp
@@ -6,7 +6,7 @@
  *  This source file is licensed as described in the file COPYING, which
  *  you should have received as part of this distribution.  The terms
  *  are also available at
- *  http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
+ *  http://www.cecill.info/licences/Licence_CeCILL_V2.1-en.txt
  *
  */
 
@@ -231,8 +231,8 @@ void Triangulator::removeDuplicateVertices(void)
 
     for (size_t i = 0; i < points.size(); i++)
     {
-        int ic = (i + 1) % points.size();
-        int icm1 = i;
+        int ic = ((int)i+1) % (int)points.size();
+        int icm1 = (int)i;
 
         Vector3d vi = points[icm1];
         Vector3d vip1 = points[ic];
